pyright

+++ cogs/Developer/owner.py
@@ -9,8 +9,10 @@ from ..utils.paginator import EmbeddedPaginator
 from ..utils.subclasses import Kana, KanaContext
 
 
-class ExtensionConverter(commands.Converter):
-    async def convert(self, ctx: KanaContext, argument: str):
+class ExtensionConverter(commands.Converter[Kana]):
+    async def convert(
+        self, ctx: commands.Context[Kana], argument: str
+    ) -> typing.Optional[typing.List[str]]:
         if argument.lower() in ("all", "*", "~"):
             return list(ctx.bot.extensions)
 
@@ -23,11 +25,11 @@ class ExtensionConverter(commands.Converter):
         return extension
 
 
-class Admin(commands.Cog):
+class Owner(commands.Cog):
     def __init__(self, bot: Kana):
         self.bot = bot
 
-    async def cog_check(self, ctx: KanaContext):
+    async def cog_check(self, ctx: commands.Context[Kana]) -> bool:
         return await self.bot.is_owner(ctx.author)
 
     @commands.command(aliases=["del"])
@@ -36,21 +38,23 @@ class Admin(commands.Cog):
         Deletes the message that the author replied to.
         """
 
-        target: discord.MessageReference = ctx.message.reference  # type: ignore
+        target: typing.Optional[discord.MessageReference] = ctx.message.reference
         if target is None:
             return await ctx.send(
                 "Reply to the message you want to delete", delete_after=5.0
             )
 
-        if target.resolved.author is not ctx.me:  # type: ignore
+        if not isinstance(target.resolved, discord.Message):
+            return
+
+        if target.resolved.author is not ctx.me:
             return await ctx.send(
                 "You can only delete messages that I sent", delete_after=5.0
             )
 
-        await target.resolved.delete()  # type: ignore
+        await target.resolved.delete()
 
     @commands.command()
-    @commands.guild_only()
     async def cleanup(self, ctx: KanaContext, limit: commands.Range[int, 1, 500] = 25):
         """
         Deletes the last `limit` amount of messages sent by the bot, if the bot has the `Manage Messages` permission, then it will also delete the command messages.
@@ -58,18 +62,24 @@ class Admin(commands.Cog):
         :param limit: The number of messages to delete, defaults to 50 if none specified.
         :type limit: int, optional
         """
+        if ctx.guild is None or isinstance(
+            ctx.channel,
+            discord.DMChannel | discord.PartialMessageable | discord.GroupChannel,
+        ):  # typehinting purposes
+            return
+
         limit += 1
-        bulk = ctx.channel.permissions_for(ctx.me).manage_messages  # type: ignore
+        bulk = ctx.channel.permissions_for(ctx.guild.me).manage_messages
 
         if bulk:
-            prefixes: typing.Tuple[str] = await self.bot.get_prefix(ctx.message)  # type: ignore
+            prefixes = await self.bot.get_prefix(ctx.message)
 
         def check(message: discord.Message):
             return message.author == ctx.me or (
                 bulk and any(message.content.startswith(prefix) for prefix in prefixes)
             )
 
-        res = await ctx.channel.purge(limit=limit, bulk=bulk, check=check)  # type: ignore
+        res = await ctx.channel.purge(limit=limit, bulk=bulk, check=check)
         if not res:
             return await ctx.send("No messages were found to cleanup.")
 
@@ -87,9 +97,10 @@ class Admin(commands.Cog):
         :type extension: str
         """
 
-        msg = []
+        msg: typing.List[str] = []
         for extension in extensions:  # type: ignore
             try:
+                extension: str
                 await self.bot.reload_extension(extension)
             except Exception as e:
                 fmt = (
@@ -97,7 +108,9 @@ class Admin(commands.Cog):
                     if (msg[-1].endswith("\n") and not msg[-1].endswith("```\n"))
                     else ""
                 )  # jank string formatting
-                msg.append(f'{fmt}\U000026a0 Could not reload `{extension}` \U000026a0 \n{to_codeblock(e, "py")}\n')  # type: ignore
+                msg.append(
+                    f'{fmt}\U000026a0 Could not reload `{extension}` \U000026a0 \n{to_codeblock(str(e), "py")}\n'
+                )
             else:
                 msg.append(f"\U0001f501 Reloaded `{extension}`\n")
 
@@ -109,5 +122,5 @@ class Admin(commands.Cog):
         ).start()
 
 
-async def setup(bot):
-    await bot.add_cog(Admin(bot))
+async def setup(bot: Kana):
+    await bot.add_cog(Owner(bot))
